pyproject.toml
[tool.poetry]
authors = ["Vasmai Avvaru <vasmaiavvaru@gmail.com>"]
description = "FastAPI microservice for telecom web app backend"
name = "telecom-web-app-backend"
version = "0.1.0-alpha"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.89.1"
pyjwt = {extras = ["crypto"], version = "^2.6.0"}
python-multipart = ">=0.0.5,<0.0.6"
sqlalchemy = "^2.0.1"
alembic = "^1.9.2"
asyncpg = "^0.27.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["dotenv", "email"], version = "^1.10.4"}
tomli = "^2.0.1"
ujson = "^5.7.0"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.0.1"
black = "^23.1.0"
coverage = "^7.1.0"
flake8 = "^6.0.0"
httpx = "^0.23.3"
isort = "^5.12.0"
pytest = "^7.2.1"
pytest-asyncio = "^0.20.3"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
pre-commit = "^3.0.4"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pytest.ini_options]
addopts = "-v"
asyncio_mode = "auto"
filterwarnings = []
markers = ["pytest.mark.asyncio"]
minversion = "6.0"
testpaths = ["app/tests"]

[tool.isort]
profile = "black"
